feat(cadastro): cria a estrutura html do cadastro

+++ src/app/cadastro/cadastro.component.ts
@@ -2,11 +2,15 @@ import { Component } from '@angular/core';
 import { Cliente } from '../model/cliente';
 import { FormsModule } from '@angular/forms';
 import { MenuComponent } from '../menu/menu.component';
+import { FontAwesomeModule } from '@fortawesome/angular-fontawesome';
+import { faEye } from '@fortawesome/free-solid-svg-icons'
+import { faEyeSlash } from '@fortawesome/free-solid-svg-icons'
+
 
 @Component({
   selector: 'app-cadastro',
   standalone: true,
-  imports: [MenuComponent, FormsModule],
+  imports: [MenuComponent, FormsModule, FontAwesomeModule],
   templateUrl: './cadastro.component.html',
   styleUrl: './cadastro.component.css'
 })
@@ -15,10 +19,17 @@ export class CadastroComponent {
     this.carregar();
   }
 
+  faEye = faEye;
+  faEyeSlash = faEyeSlash;
+
 
   public cliente:Cliente = new Cliente();
   public mensagem:String = "";
+  public mensagemErro:String = "";
   public cadastrarBtnText:String = "";
+  public confirmarSenhaValor:String = ""
+  public mostraSenha: boolean = false;
+  public mostraConfirmSenha: boolean = false;
 
   public carregar(){
     let json:any = localStorage.getItem("cliente");
@@ -28,14 +39,31 @@ export class CadastroComponent {
       this.cadastrarBtnText = "Atualizar dados";
     }
     else{
-      this.mensagem = "Insira os dados para um novo cadastro";
-      this.cadastrarBtnText = "Criar uma nova conta";
+      this.mensagem = "Criar uma nova conta";
+      this.cadastrarBtnText = "Criar conta";
     }
   }
 
   public salvarCadastro(){
-    localStorage.setItem("clienteMemoria", JSON.stringify(this.cliente));
-    this.mensagem = "Cadastro atualizado com sucesso!";
-    window.location.href="./login";
+    if(this.cliente.senha == this.confirmarSenhaValor){
+      localStorage.setItem("clienteMemoria", JSON.stringify(this.cliente));
+      this.mensagem = "Cadastro atualizado com sucesso!";
+      window.location.href="./login";
+    }
+    else{
+      this.mensagemErro = "As senhas informadas são diferentes!";
+    }
+  }
+
+
+  public alternaSenha() {
+    this.mostraSenha = !this.mostraSenha;
+  }
+
+  public alternaConfirmSenha() {
+    this.mostraConfirmSenha = !this.mostraConfirmSenha;
   }
 }
+  
+
+
